Avoid exec:exec hack
To be able to load the native library, we introduced a hack to fork
a new Java process with a hand-crafted java.library.path. But that
system was not flexible enough: it only worked for the same platform
as it was developed on. So let's not go the NarSystem route but do
something slightly more clever.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>