NarHelper: support launching from inside .jar files, too
Previously, we punted on finding the native library unless the
corresponding nar.properties were available via a plain file: URL. Now
we do better, we handle jar:file: URLs, too.

We still do not support unpacking the native library from a .nar file,
but expect the library to be available where the nar-maven-plugin
compiled it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>